Implement the new patient finder, patient details and surgeries page for the inpatient module
