+++ interface/modules/custom_modules/oe-inpatient-module/public/components/templates/msv-surgeries.php
@@ -1,0 +1,17 @@
+<?php
+ob_start();
+?>
+
+<div>
+    <?php
+    include_once __DIR__ . "/../organisms/msv-surgeries-content.php";
+    ?>
+</div>
+
+<?php
+
+$pageContent = ob_get_clean();
+$content = $pageContent; // Pass the captured content to the layout
+
+include __DIR__ . '/../layouts/msv-inpatient-layout.php';
+?>
