+++ interface/modules/custom_modules/oe-inpatient-module/public/dashboard.php
@@ -1,0 +1,3 @@
+<?php
+ include_once __DIR__ . "/components/templates/msv-inpatient-dashboard.php";
+?>
